+++ src/MenuPanel.php
@@ -1,6 +1,8 @@
 <?php
 namespace GCWorld\Menu;
 
+use GCWorld\Menu\Core\Twig;
+
 /**
  * Class MenuPanel
  */
@@ -47,25 +49,13 @@ class MenuPanel
      */
     public function returnPanel(bool $slim = false): string
     {
-        $out = '';
-        foreach ($this->blocks as $block) {
-            /** @var MenuBlock $obj */
-            $obj = $block['obj'];
+        $columnWidth = empty($this->blocks) ? 12 : floor(12/count($this->blocks));
 
-            $out .= '<div class="col-sm-'.floor(12/count($this->blocks)).'">';
-            if (!$slim && $block['name'] != 'Spacer' && $obj->wrap) {
-                $out .= '<div class="panel panel-sitespecific">';
-                $out .= '<div class="panel-heading"><div class="panel-title">'.$block['name'].'</div></div>';
-                $out .= '<div class="panel-body">';
-            }
-            $out .= $obj->returnBlock();
-            if (!$slim && $block['name'] != 'Spacer' && $obj->wrap) {
-                $out .= '</div>';
-                $out .= '</div>';
-            }
-            $out .= '</div>';
-        }
-        return $out;
+        return Twig::render('@GCMenu/menu_panel.twig', [
+            'blocks'       => $this->blocks,
+            'slim'         => $slim,
+            'column_width' => $columnWidth,
+        ]);
     }
 
     /**