+++ tests/Core/TwigTest.php
@@ -37,4 +37,28 @@ final class TwigTest extends TestCase
         );
         self::assertSame(Twig::getLoader(), $twig->getLoader());
     }
+
+    public function testEveryRendererTemplateIsPackaged(): void
+    {
+        $templateDirectory = dirname(__DIR__, 2).DIRECTORY_SEPARATOR.'twig';
+        $templates = [
+            'menu.twig',
+            'menu_elements.twig',
+            'drop_down_normal.twig',
+            'drop_down_wide.twig',
+            'drop_down_notices.twig',
+            'menu_panel.twig',
+            'menu_block.twig',
+            'components/drop_down_notice_item.twig',
+            'panel_elements/link.twig',
+        ];
+
+        foreach ($templates as $template) {
+            self::assertFileExists($templateDirectory.DIRECTORY_SEPARATOR.$template);
+            self::assertSame(
+                '@GCMenu/'.$template,
+                Twig::get()->load('@GCMenu/'.$template)->getTemplateName(),
+            );
+        }
+    }
 }